+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+# 0.7.2
+
+* fix bug in debug log message.
+
+# 0.7.1
+
+* Fix get_list method not returning uuid and status from JSON response
+* Log to debug instead of error where exceptions are already raised.
+* Add raw server response to debug log messages.
+* Update docstrings
+
 # 0.7.0
 
 * **New API method:** `retrieve_new_access_token` retrieves a new access token and updates authorization headers. Time till expiration of the access token is stored in the property `expires_in` . ([tr4nt0r](https://github.com/tr4nt0r))